web: Move storage proposal actions to a dialog
With some skipped tests.
