+++ web/src/components/storage/index.js
@@ -24,8 +24,8 @@ export { default as ProposalPageMenu } from "./ProposalPageMenu";
 export { default as ProposalSettingsSection } from "./ProposalSettingsSection";
 export { default as ProposalSpacePolicyField } from "./ProposalSpacePolicyField";
 export { default as ProposalDeviceSection } from "./ProposalDeviceSection";
-export { default as ProposalActionsSection } from "./ProposalActionsSection";
 export { default as ProposalTransactionalInfo } from "./ProposalTransactionalInfo";
+export { default as ProposalActionsDialog } from "./ProposalActionsDialog";
 export { default as ProposalResultSection } from "./ProposalResultSection";
 export { default as ProposalVolumes } from "./ProposalVolumes";
 export { default as DASDPage } from "./DASDPage";
